+++ app/javascript/vue/tasks/otu/browse/components/BiologicalAssociations.vue
@@ -3,6 +3,7 @@
     :status="status"
     :title="title"
     :spinner="loadState.biologicalAssociations"
+    menu
     @menu="setModalView(true)">
     <a name="biological-associations"/>
     <table class="full_width">