+++ app/javascript/vue/tasks/otu/browse/components/NomenclatureHistory.vue
@@ -3,6 +3,7 @@
     :status="status"
     :title="title"
     :spinner="isLoading"
+    menu
     @menu="showModal = true">
     <a name="timeline"/>
     <div class="switch-radio separate-top separate-bottom">